+++ deno/v0.1/mod.ts
@@ -2,3 +2,5 @@ export type {
   BroadcasterInterface,
   MessageInterface,
 } from "./type_flyweight/router.ts";
+
+export { Router } from "./router/router.ts";
