--- docs/deno/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["dom", "esnext"]
-  }
-}
