--- dom/v0.1/deps.ts
@@ -1,4 +1,0 @@
-export type {
-  BroadcasterInterface,
-  MessageInterface,
-} from "../../deno/v0.1/mod.ts";
